requirements.txt
altgraph==0.17.2
appdirs==1.4.3
apturl==0.5.2
auto-py-to-exe==2.10.1
bcrypt==3.1.7
blinker==1.4
bottle==0.12.19
bottle-websocket==0.2.9
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==2.2.12
django-heroku==0.3.1
djangorestframework==3.12.4
duplicity==0.8.12.0
Eel==0.12.4
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pipenv==11.9.0
protobuf==3.6.1
psycopg2==2.9.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.2.4
ssh-import-id==5.10
systemd-python==234
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
whichcraft==0.6.1
whitenoise==5.3.0
xkit==0.0.0
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
